Split Integration tests into base class
